refactor(filters -> processors): merge filters and middlewares into "Modifiers" feat(middlewares): implement middlewares as a part of modifiers feat(pocket): provide a redis client for storage, named "pocket"
BREAKING CHANGE: filters renamed to modifiers, now accessible as a property on bot.

Modifiers:
  - Proccessors: includes preprocessors and postprocessors used to modify a function's
                 arguments and return value
  - Middlewares: used to change a function's behavior, usually triggering an action
                 right before executing the final call in a function.

Middlewares are passed a single argument `context` which includes the information
available to middleware. Middlewares should return a promise, resolving in case the middleware
decides the chain should continue or reject to break the chain of middlewares.
If all middlewares pass, the actual function runs.

Pocket: A redis storage used by bots. We use the `redis` module for this, with an addition
of promisifying all methods using `bluebird`.